pyproject.toml
[project]
name = "olist-analytics-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-duckdb>=1.10.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyarrow>=22.0.0",
    "rich>=14.2.0",
    "streamlit>=1.52.2",
]
